1. Moving reducers to their folder
