+++ interview-scheduler/src/actions.js
@@ -1,0 +1,4 @@
+export const testScoreEdit = (score) => ({
+    TYPE: "TEST_SCORE",
+    payload: score
+});
