+++ interview-scheduler/src/reducers/users-reducer.js
@@ -1,4 +1,4 @@
-import MockData from './mock-data.json';
+import MockData from '../mock-data.json';
 
 export default function () {
     return MockData.data;
