+++ interview-scheduler/src/store/store.js
@@ -1,5 +1,5 @@
 import { createStore, applyMiddleware } from 'redux';
-import rootReducer from '../reducers';
+import rootReducer from '../reducers/root-reducer';
 import logger from 'redux-logger'
 
 const store = createStore(rootReducer, applyMiddleware(logger));
